+++ GroovyfxGriffonPlugin.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2012 the original author or authors.
+ * Copyright 2010-2013 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the 'License');
  * you may not use this file except in compliance with the License.
@@ -18,8 +18,8 @@
  * @author Dean Iverson
  */
 class GroovyfxGriffonPlugin {
-    String version = '0.8'
-    String griffonVersion = '1.1.0 > *'
+    String version = '0.9.0'
+    String griffonVersion = '1.2.0 > *'
     Map dependsOn = [:]
     List pluginIncludes = []
     String license = 'Apache Software License 2.0'
@@ -47,26 +47,28 @@ class GroovyfxGriffonPlugin {
     ]
     String title = 'Griffon GroovyFX Plugin'
     String description = '''
-Provides integration with [GroovyFX][1].  GroovyFX allows you to use the [JavaFX][2] API to write your Griffon
-views.
+Provides integration with [GroovyFX][1]. GroovyFX allows you to use the [JavaFX][2]
+API to write your Griffon views.
 
-*Note:* If you want to create a Griffon JavaFX application, install the JavaFX plugin instead.  The JavaFX plugin
-will automatically pull this plugin in as a dependency.  This plugin should only be installed independently if
-you want to add JavaFX views to your existing Swing application.
+*Note:* If you want to create a Griffon JavaFX application, install the JavaFX plugin
+instead.  The JavaFX plugin will automatically pull this plugin in as a dependency.
+This plugin should only be installed independently if you want to add JavaFX views
+to your existing Swing application.
 
 Usage
 -----
 
-This version of the plugin uses GroovyFX 0.2, which works with JavaFX 2.1.  The JavaFX 2.1 SDK can be downloaded
-from [javafx.com][3].  Once the SDK is installed, the JAVAFX_HOME environment variable should be set to point to
-the root directory of the JavaFX SDK installation
+This version of the plugin uses GroovyFX 0.3.1, which works with JavaFX 2.2. The
+JavaFX 2.2 SDK can be downloaded from [javafx.com][3].  Once the SDK is installed,
+the JAVAFX_HOME environment variable should be set to point to the root directory
+of the JavaFX SDK installation
 
-The JavaFX 2.1 SDK is also bundled with the Java 7u4 (or later) JDK.  In this case, you can just set your
-JAVA_HOME environment variable to point to the JDK installation and the plugin will use the bundled JavaFX
-SDK.
+The JavaFX 2.2 SDK is also bundled with the Java 7u13 (or later) JDK. In this
+case, you can just set your JAVA_HOME environment variable to point to the JDK
+installation and the plugin will use the bundled JavaFX SDK.
 
-Once installed, you can use any of the nodes provided by GroovyFX in your views.  GroovyFX documentation can be
-found [here][4].
+Once installed, you can use any of the nodes provided by GroovyFX in your views.
+GroovyFX documentation can be found [here][4].
 
 ### Example