+++ griffon-app/conf/BuildConfig.groovy
@@ -2,11 +2,8 @@ griffon.project.dependency.resolution = {
     inherits("global")
     log "warn"
     repositories {
-        griffonPlugins()
         griffonHome()
-        griffonCentral()
         mavenCentral()
-        mavenLocal()
 //        mavenRepo "https://oss.sonatype.org/content/groups/public"
         // pluginDirPath is only available when installed
 //        String basePath = pluginDirPath? "${pluginDirPath}/" : ''
@@ -17,17 +14,7 @@ griffon.project.dependency.resolution = {
     }
 }
 
-griffon {
-    doc {
-        logo = '<a href="http://griffon.codehaus.org" target="_blank"><img alt="The Griffon Framework" src="../img/griffon.png" border="0"/></a>'
-        sponsorLogo = "<br/>"
-        footer = "<br/><br/>Made with Griffon (@griffon.version@)"
-    }
-}
-
 log4j = {
-    // Example of changing the log pattern for the default console
-    // appender:
     appenders {
         console name: 'stdout', layout: pattern(conversionPattern: '%d [%t] %-5p %c - %m%n')
     }